build: code cleanup
